backend/pyproject.toml
[project]
name = "pfms-backend"
version = "0.1.0"
dependencies = ["fastapi>=0.115", "uvicorn[standard]>=0.30", "pydantic>=2.7", "sqlalchemy>=2.0", "python-dotenv>=1.0", "PyJWT>=2.0", "passlib[bcrypt]>=1.7", "python-multipart>=0.0.5"]

[tool.ruff]
line-length = 100
